Convert last_n_errors_logger to ts.
Add jsdoc. Add types.
