+++ server/control.js
@@ -16,7 +16,7 @@ limitations under the License.
 import * as wallGeometry from './util/wall_geometry.ts';
 import * as time from './util/time.ts';
 import * as network from './network/network.js';
-import {getErrors} from './util/last_n_errors_logger.js';
+import {getErrors} from './util/last_n_errors_logger.ts';
 import {loadAllModules} from './playlist/playlist_loader.js';
 import {WSS} from './network/websocket.ts';
 import {easyLog} from '../lib/log.js';
