+++ server/server.ts
@@ -29,7 +29,7 @@ import {
   loadPlaylistFromFile,
 } from "./playlist/playlist_loader.js";
 import { makeConsoleLogger } from "../lib/console_logger.js";
-import { captureLog } from "./util/last_n_errors_logger.js";
+import { captureLog } from "./util/last_n_errors_logger.ts";
 import { addLogger, easyLog } from "../lib/log.js";
 import * as colors from "https://deno.land/std@0.123.0/fmt/colors.ts";
 import { now } from "./util/time.ts";
